Implementing /number-of-files endpoint
This endpoint allows getting the total number of files
in the storage service. It simply reads an atomic long
variable that keeps track of the total number of files
stored.

Such an approach should not introduce too much overhead
since the only bottleneck for the file access operations
remains the update of this variable (everything else can
still happen in parallel).

+++ src/main/java/com/andreitraistaru/filestorage/dto/NumberOfFilesDTO.java
@@ -1,13 +1,13 @@
 package com.andreitraistaru.filestorage.dto;
 
 public class NumberOfFilesDTO {
-    private int numberOfFiles;
+    private long numberOfFiles;
 
-    public int getNumberOfFiles() {
+    public long getNumberOfFiles() {
         return numberOfFiles;
     }
 
-    public void setNumberOfFiles(int numberOfFiles) {
+    public void setNumberOfFiles(long numberOfFiles) {
         this.numberOfFiles = numberOfFiles;
     }
 }
